+++ demo/src/stories/button.stories.jsx
@@ -25,7 +25,7 @@ export const Button5 = () => <Button style={{ backgroundColor: 'red', color: 'ho
 Button5.storyName = 'Button with insufficient color contrast but skipped (should pass)';
 Button5.parameters = {
   axe: {
-    disabled: true,
+    skip: true,
   },
 };